+++ rs/moq-relay/src/lib.rs
@@ -11,10 +11,8 @@ mod auth;
 mod cluster;
 mod config;
 mod connection;
-mod health;
 mod http_client;
 mod stats;
-mod watch;
 mod web;
 #[cfg(feature = "websocket")]
 mod websocket;
@@ -27,6 +25,5 @@ pub use auth::*;
 pub use cluster::*;
 pub use config::*;
 pub use connection::*;
-pub use health::*;
 pub use stats::*;
 pub use web::*;